STEP3 수정
